requirements.txt
Django==3.2.4
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-storages==1.11.1
fpdf==1.7.2